Reorganize source tree
