+++ src/index.ts
@@ -2,16 +2,11 @@
  * @packageDocumentation
  * @module Module @proc7ts/context-values
  */
-export * from './context-builder';
-export * from './context-key';
 export * from './context-key-error';
+export * from './context-request';
 export * from './context-ref';
-export * from './context-registry';
-export * from './context-seeder';
 export * from './context-supply';
-export * from './context-value-spec';
 export * from './context-values';
-export * from './iterative-context-key';
-export * from './multi-context-key';
-export * from './simple-context-key';
-export * from './single-context-key';
+export * from './registry';
+export * from './singleton';
+export * from './key';
