+++ src/singleton/iterative-context-key.ts
@@ -8,10 +8,10 @@ import {
   PushIterable,
   valueIt,
 } from '@proc7ts/push-iterator';
-import { ContextKey, ContextSeedKey } from './context-key';
-import type { ContextSeeder } from './context-seeder';
-import type { ContextValueProvider } from './context-value-spec';
-import type { ContextValues } from './context-values';
+import type { ContextValues } from '../context-values';
+import type { ContextSeeder } from '../key';
+import { ContextKey, ContextSeedKey } from '../key';
+import type { ContextValueProvider } from '../registry';
 
 /**
  * @internal
