+++ src/singleton/single-context-key.ts
@@ -1,6 +1,6 @@
 import { noop } from '@proc7ts/primitives';
-import type { ContextKey, ContextKeyDefault, ContextSeedKey, ContextValueSlot } from './context-key';
-import type { ContextRef } from './context-ref';
+import type { ContextRef } from '../context-ref';
+import type { ContextKey, ContextKeyDefault, ContextSeedKey, ContextValueSlot } from '../key';
 import { SimpleContextKey } from './simple-context-key';
 
 /**
