+++ src/updatable/fn-context-key.spec.ts
@@ -1,8 +1,8 @@
 import { Supply } from '@proc7ts/primitives';
 import { ContextKeyError } from '../context-key-error';
-import { ContextRegistry } from '../context-registry';
 import { ContextSupply } from '../context-supply';
 import type { ContextValues } from '../context-values';
+import { ContextRegistry } from '../registry';
 import { FnContextKey } from './fn-context-key';
 
 describe('FnContextKey', () => {
