+++ src/updatable/index.ts
@@ -3,9 +3,8 @@
  * @module Module @proc7ts/context-values/updatable
  */
 export * from './context-destroyed';
-export * from './context-module';
-export * from './context-module-dependency-error';
 export * from './context-up-key';
 export * from './fn-context-key';
 export * from './multi-context-up-key';
 export * from './single-context-up-key';
+export * from './modules';
