+++ src/updatable/modules/context-module-usage.impl.ts
@@ -9,10 +9,10 @@ import {
   ValueTracker,
 } from '@proc7ts/fun-events';
 import { neverSupply, Supply, SupplyPeer, valueProvider } from '@proc7ts/primitives';
-import type { ContextRequest } from '../context-ref';
-import type { ContextRegistry } from '../context-registry';
-import { ContextSupply } from '../context-supply';
-import type { ContextValues } from '../context-values';
+import type { ContextRequest } from '../../context-request';
+import { ContextSupply } from '../../context-supply';
+import type { ContextValues } from '../../context-values';
+import type { ContextRegistry } from '../../registry';
 import type { ContextModule } from './context-module';
 
 /**
