+++ src/updatable/modules/context-module.impl.ts
@@ -1,11 +1,11 @@
 import type { OnEvent } from '@proc7ts/fun-events';
 import { isDefined, noop, setOfElements, Supply, valueProvider } from '@proc7ts/primitives';
 import { itsElements, valueIt } from '@proc7ts/push-iterator';
-import type { ContextRegistry } from '../context-registry';
+import type { ContextRegistry } from '../../registry';
+import type { ContextUpKey } from '../context-up-key';
 import type { ContextModule } from './context-module';
 import { ContextModuleDependencyError } from './context-module-dependency-error';
 import { ContextModuleKey } from './context-module-key.impl';
-import type { ContextUpKey } from './context-up-key';
 
 /**
  * @internal
