+++ src/updatable/modules/index.ts
@@ -1,0 +1,2 @@
+export * from './context-module';
+export * from './context-module-dependency-error';
